проекты/bot + parser cian/tgbot/requirements.txt
python-telegram-bot[job-queue]>=20.0
gigachat>=0.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
requests>=2.31.0
selenium>=4.10.0
pandas>=2.0.0













